Fix: kotlin compiler bug requires cast to inline members
The issue is that if the derived class instance isn't casted, the kotlin
compiler will fail with a weird, cryptic error message saying that it
can't inline the said members because it can't resolve the member
definitions, and it isn't even obvious that the instance of the derived
class must simply be casted first. This seems to be a bug in the kotlin
compiler.